MQDT

Documentation for MQDT.

MQDT.BasisArray

MQDT.BasisState

MQDT.Channels

MQDT.DataBaseArray

MQDT.DataBaseState

MQDT.EigenStates

MQDT.Model

MQDT.Parameters

MQDT.QuantumNumbers

MQDT.fModel

MQDT.fjChannels

MQDT.fjQuantumNumbers

MQDT.jjChannels

MQDT.jjQuantumNumbers

MQDT.kModel

MQDT.lsChannels

MQDT.lsQuantumNumbers

LinearAlgebra.nullspace

MQDT.G1

MQDT.G2

MQDT.G3

MQDT.G4

MQDT.G5

MQDT.G6

MQDT.G7

MQDT.G8

MQDT.angular_matrix_cached

MQDT.angular_moment

MQDT.basisarray

MQDT.beta

MQDT.couple

MQDT.databasearray

MQDT.eigenenergies

MQDT.eigenstates

MQDT.element_ic

MQDT.element_lc

MQDT.element_lr

MQDT.element_sc

MQDT.element_sr

MQDT.epsilon

MQDT.get_rydberg_state_cached

MQDT.good_quantum_number

MQDT.kbar

MQDT.kmat

MQDT.magnetic_matrix_cached

MQDT.magneton

MQDT.mmat

MQDT.mroots

MQDT.multipole_moments

MQDT.nu

MQDT.parity

MQDT.radial_matrix

MQDT.radial_moment_cached

MQDT.rot

MQDT.square_brakets

MQDT.state_data

MQDT.theta

MQDT.theta_rr

MQDT.transform

MQDT.unique_parity

MQDT.Λ

MQDT.BasisArray — Type

See also basisarray, BasisState

BasisArray(states::Vector{BasisState})

Type to store a basis. A basis is a list of BasisState types, typically corresponding to different Rydberg series. BasisArray is generated by the basisarray function using lists of Model and EigenStates.

MQDT.BasisState — Type

See also basisarray, BasisArray, EigenStates

BasisState(
    species::Symbol,
    energy::Float64,
    parity::Int,
    f::Float64,
    nu::Vector{Float64},
    lr::Vector{Int},
    coeff::Vector{Float64},
    channels::Channels,
    term::String,
    lead::Float64,
    L::Float64,
    S::Float64,
)

Type to store all relevant information of multi-channel bound states for a given Rydberg series. BasisState is generated by the basisarray function using a Model and EigenStates.

MQDT.Channels — Type

See also lsChannels, jjChannels, fjChannels

Abstract type to store the quantum numbers of an MQDT channel representation.

MQDT.DataBaseArray — Type

See also DataBaseState, databasearray, basisarray, BasisArray

DataBaseArray(states::Vector{DataBaseState})

Type to store a bound state data as a list of DataBaseState types as used by the PAIRINTERACTION software. DataBaseArray is generated by the databasearray function using lists of Model and EigenStates.

MQDT.DataBaseState — Type

See also databasearray, DataBaseArray, EigenStates, BasisState

DataBaseState(
    nu::Float64,
    parity::Int,
    f::Float64,
    nui::Vector{Float64},
    nui_all::Vector{Float64},
    ai::Vector{Float64},
    ai_all::Vector{Float64},
    S::Vector{Float64},
    L::Vector{Float64},
    J::Vector{Float64},
    lr::Vector{Int},
    Jr::Vector{Float64},
    transform::Matrix{Float64},
    term::String,
    lead::Float64,
    neg::Float64,
)

Type to store all relevant to the PAIRINTERACTION database for a given Rydberg series. DataBaseState is generated by the databasearray function using a Model and EigenStates.

MQDT.EigenStates — Type

See also eigenstates, BasisState

EigenStates(n::Vector{Float64}, nu::Matrix{Float64}, a::Matrix{Float64})

Type to store multi-channel bound states. n is a global energy reference to a global threshold specified in the Parameters type. nu is the principal quantum number in each channel with respect to the channel-specific threshold. a is the coefficient of each channel contributing to the bound state. EigenStates is populated by the eigenstates function.

MQDT.Model — Type

See also fModel, kModel

Abstract type to store MQDT models.

MQDT.Parameters — Type

Parameters(species::Symbol, mass::Float64, spin::Float64, rydberg::Float64, threshold::Float64, hyperfine::Float64, dipole::Float64)

Type to store relevant parameters for each atomic element. A corresponding file in the parameters folder of the mqdt module initializes this type.

Examples

julia> PARA = Parameters(
           :Yb174,
           1822.88848628*173.9388621, # nuclear mass
           0, # nuclear spin
           109736.9695858, # Rydberg constant in 1/cm
           50443.070393, # lowest ionization threshold in 1/cm
           0, # hyperfine constant in 1/cm
           2.1, # nuclear dipole
       );

MQDT.QuantumNumbers — Type

See also lsQuantumNumbers, jjQuantumNumbers, fjQuantumNumbers

Abstract type to store the quantum numbers of an MQDT channel representation.

MQDT.fModel — Type

See also kModel

fModel(
    species::Symbol,
    name::String,
    size::Int,
    terms::Vector{String},
    core::Vector{Bool},
    thresholds::Vector{Float64},
    defects::Matrix{Float64},
    mixing::Vector{String},
    angles::Matrix{Float64},
    inner_channels::Channels
    outer_channels::Channels
    unitary::Matrix{Float64}
)

Type to store MQDT models inspired by [PRX 15, 011009 (2025)] using sparse K matriced and frame transformation. Contains all relevant parameters to form the K matrix and calculate the spectrum for a specific Rydberg series.

Examples

julia> RYDBERG_S0 = fModel(
           :Yb174,
           "S J=0, ν > 2", # fit for states 6s7s upward [Phys. Rev. X 15, 011009 (2025)]
           6,
           ["6sns 1S0", "4f13 5d 6snl a", "6pnp 1S0", "4f13 5d 6snl b", "6pnp 3P0", "4f13 5d 6snl c"],
           Bool[1, 0, 1, 0, 1, 0],
           [50443.070393, 83967.7, 80835.39, 83967.7, 77504.98, 83967.7],
           [
               0.355097325 0.278368431;
               0.204537279 0;
               0.116394359 0;
               0.295432196 0;
               0.25765161 0;
               0.155807042 0
           ],
           ["12", "13", "14", "34", "35", "16"],
           [0.12654859 0; 0.30010744 0; 0.05703381 0; 0.11439805 0; 0.09864375 0; 0.14248210 0],
           lsChannels([
               lsQuantumNumbers(0.5, 0, 0, 0, 0, 0),
               lsQuantumNumbers(0.5, 0, 1, 1, 0, 0),
               lsQuantumNumbers(0.5, 1, 1, 1, 1, 0),
           ]),
           jjChannels([
               jjQuantumNumbers(0.5, 0, 0.5, 0, 0.5, 0),
               jjQuantumNumbers(0.5, 1, 0.5, 1, 1.5, 0),
               jjQuantumNumbers(0.5, 1, 0.5, 1, 0.5, 0),
           ]),
           [
               1 0 0 0 0 0;
               0 1 0 0 0 0;
               0 0 -sqrt(2/3) 0 sqrt(1/3) 0;
               0 0 0 1 0 0;
               0 0 sqrt(1/3) 0 sqrt(2/3) 0;
               0 0 0 0 0 1
           ],
       );

MQDT.fjChannels — Type

See also lsChannels, jjChannels

fjChannels(i::Vector{fjQuantumNumbers})

Type to store the quantum numbers of an MQDT channel representation, here in the "fj" (sometimes called "fragmentation") coupling scheme

MQDT.fjQuantumNumbers — Type

See also lsQuantumNumbers, jjQuantumNumbers

fjQuantumNumbers(sc::Float64, lc::Int, Jc::Float64, Fc::Float64, lr::Int, Jr::Float64, F::Float64)

Type to store the quantum numbers of a single MQDT channel in the "fj" (sometimes called "fragmentation") coupling scheme, i.e. core angular momentum quantum numbers are coupled to the nuclear spin to give a core hyperfine configuration (Fc) which is coupled to the total Rydberg angular momentum (Jr).

MQDT.jjChannels — Type

See also lsChannels, fjChannels

jjChannels(i::Vector{jjQuantumNumbers})

Type to store the quantum numbers of an MQDT channel representation, here in the "jj" coupling scheme, i.e. core and Rydberg angular momentum (AM) quantum numbers are coupled separately to give a total core AM (Jc) and total Rydberg AM (Jr).

MQDT.jjQuantumNumbers — Type

See also lsQuantumNumbers, fjQuantumNumbers

jjQuantumNumbers(lc::Int, Jc::Float64, lr::Int, Jr::Float64, J::Float64)

Type to store the quantum numbers of a single MQDT channel in the "jj" coupling scheme, i.e. core and Rydberg angular momentum (AM) quantum numbers are coupled separately to give a total core AM (Jc) and total Rydberg AM (Jr).

MQDT.kModel — Type

See also fModel

kModel(
    species::Symbol
    name::String
    size::Int
    terms::Vector{String}
    jjscheme::Vector{Bool}
    lschannels::lsChannels
    jjchannels::jjChannels
    thresholds::Vector{Float64}
    K0::Matrix{Float64}
    K1::Vector{Float64}
)

Type to store MQDT models inspired by [JPB 47, 155001 (2014)] using dense, energy-dependent K matrices. Contains all relevant parameters to form the K matrix and calculate the spectrum for a specific Rydberg series.

Examples

julia> KMODEL_S0 = kModel(
           :Sr88,
           "1S0",
           3,
           ["(5s1/2)(ns1/2)", "(4d5/2)(nd5/2)", "(4d3/2)(nd3/2)"],
           Bool[1, 1, 1],
           lsChannels([
               lsQuantumNumbers(0.5, 0, 0, 0, 0, 0),
               lsQuantumNumbers(0.5, 0, 2, 2, 0, 0),
               lsQuantumNumbers(0.5, 0, 2, 2, 0, 0),
           ]),
           jjChannels([
               jjQuantumNumbers(0.5, 0, 0.5, 0, 0.5, 0),
               jjQuantumNumbers(0.5, 2, 2.5, 2, 2.5, 0),
               jjQuantumNumbers(0.5, 2, 1.5, 2, 1.5, 0),
           ]),
           [45932.2002, 60768.43, 60488.09],
           [
               1.051261 0.3759864 -0.02365485;
               0.3759864 -0.6400925 -0.0002063825;
               -0.02365485 -0.0002063825 3.009087
           ],
           [0.8763911, 0.4042584, 17.22631],
       );

MQDT.lsChannels — Type

See also jjChannels, fjChannels

lsChannels(i::Vector{lsQuantumNumbers})

Type to store the quantum numbers of an MQDT channel representation, here in the "LS" coupling scheme, i.e. core and Rydberg angular momentum quantum numbers are coupled individually to give a total spin (S) and total orbital angular momentum (L).

MQDT.lsQuantumNumbers — Type

See also jjQuantumNumbers, fjQuantumNumbers

lsQuantumNumbers(S::Float64, lc::Int, lr::Int, L::Int, J::Float64)

Type to store the quantum numbers of a single MQDT channel in the "LS" coupling scheme, i.e. core and Rydberg angular momentum quantum numbers are coupled individually to give a total spin (S) and total orbital angular momentum (L).

LinearAlgebra.nullspace — Method

nullspace(n::Vector{Float64}, M::Matrix)

Returns the nullspace of the M matrix, which corresponds to the channel coefficents for an MQDT bound state.

MQDT.G1 — Method

G1(q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Auxiliary function to calculate reduced matrix elements for the magnetic moment (Robicheaux2018 Eq. 25)

MQDT.G2 — Method

G2(q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Auxiliary function to calculate reduced matrix elements for the magnetic moment (Robicheaux2018 Eq. 25)

MQDT.G3 — Method

G3(q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Auxiliary function to calculate reduced matrix elements for the magnetic moment (Robicheaux2018 Eq. 25)

MQDT.G4 — Method

G4(q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Auxiliary function to calculate reduced matrix elements for the magnetic moment (Robicheaux2018 Eq. 25)

MQDT.G5 — Method

G5(q1::fjQuantumNumbers, q2::fjQuantumNumbers, ic)

Auxiliary function to calculate reduced matrix elements for the magnetic moment (Robicheaux2018 Eq. 25)

MQDT.G6 — Method

G6(q1::fjQuantumNumbers, q2::fjQuantumNumbers, ic)

Auxiliary function to calculate reduced matrix elements for the magnetic moment (Robicheaux2018 Eq. 25)

MQDT.G7 — Method

G7(q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Auxiliary function to calculate reduced matrix elements for the magnetic moment (Robicheaux2018 Eq. 25)

MQDT.G8 — Method

G8(q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Auxiliary function to calculate reduced matrix elements for the magnetic moment (Robicheaux2018 Eq. 25)

MQDT.angular_matrix_cached — Method

See also angular_moment

angular_matrix_cached(k::Int, k1::Channels, k2::Channels)

Iterates angular_moment over Channels, returns a matrix. This function is cached using the LRUCache package.

MQDT.angular_moment — Method

See also angular_matrix_cached

angular_moment(k, q1::fjQuantumNumbers, q2::fjQuantumNumbers)
angular_moment(k, q1::jjQuantumNumbers, q2::jjQuantumNumbers)
angular_moment(k, q1::lsQuantumNumbers, q2::lsQuantumNumbers)

Returns the angular matrix elements (i.e. an analytically evaluated integral of spherical harmonics). Formula is found in Robicheaux2018 Eq. 20 and in Vaillant2014 Eq. C3

MQDT.basisarray — Method

basisarray(T::EigenStates, M::fModel)
basisarray(T::Vector{EigenStates}, M::Vector{fModel})

Function that generates all relevant bound-state data. Returns a BasisArray, which is a list of BasisState instances.

MQDT.beta — Method

beta(N, L) = π .* (N .- L)

MQDT.couple — Method

couple(C::String)

From a string of type "ij", return a touple of indices (i, j). With this method, maximum for j is 10.

Examples

MQDT.couple("710")

# output

(7, 10)

MQDT.databasearray — Method

dbstates(T::EigenStates, M::fModel)
dbstates(T::Vector{EigenStates}, M::Vector{fModel})

Function that generates all relevant bound-state data specifically for the PAIRINTERACTION database.

MQDT.eigenenergies — Method

See also mroots

eigenenergies(N1::Number, N2::Number, M::Model, P::Parameters)

Function that returns both the roots of the M matrix as well as the channel-dependent principal quantum numbers.

MQDT.eigenstates — Method

See also eigenenergies

eigenstates(N1::Number, N2::Number, M::Model, P::Parameters)

Function that returns the bound states corresponding to an MQDT model in the form of an instance of EigenStates, which contains the global reference principal quantum number, the channel-dependent principal quantum numbers, as well as the channel coefficients.

MQDT.element_ic — Method

element_ic(ic, q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Returns the reduced matrix element of core nuclear spin (Robicheaux2018 Eq. 24)

MQDT.element_lc — Method

element_lc(ic, q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Returns the reduced matrix element of core orbital anugular momentum (Robicheaux2018 Eq. 24)

MQDT.element_lr — Method

element_lr(q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Returns the reduced matrix element of Rydberg orbital anugular momentum (Robicheaux2018 Eq. 24)

MQDT.element_sc — Method

element_sc(ic, q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Returns the reduced matrix element of core spin (Robicheaux2018 Eq. 24)

MQDT.element_sr — Method

element_sr(q1::fjQuantumNumbers, q2::fjQuantumNumbers)

Returns the reduced matrix element of Rydberg spin (Robicheaux2018 Eq. 24)

MQDT.epsilon — Method

epsilon(N, T, R)
epsilon(N, P::Parameters)

Given a principal quantum number, return the energy, depending on the ionization thresholds and the Rydberg constant.

Examples

ϵ = MQDT.epsilon(60, 50443.217463, 109736.9635066)

# output

50412.734973137056

MQDT.get_rydberg_state_cached — Method

get_rydberg_state_cached(species::Symbol, nu::Float64, l::Int64)

Get a rydberg state from the ryd-numerov package and calculate its wavefunction. This function is cached using the LRUCache package.

MQDT.good_quantum_number — Method

good_quantum_number(T::lsChannels)
good_quantum_number(T::jjChannels)
good_quantum_number(T::fjChannels)

Given a ls or jj (fj) channel representation, this function checks whether each channel has the same total (hyperfine) angular momentum J (F) and, if true, returns it.

MQDT.kbar — Method

kbar(M::Number) = tan(π * M)
kbar(M::Vector{Float64}) = diagm(kbar.(M))
kbar(N::Number, M::fModel, P::Parameters)

Function that returns the K matrix in the close-coupling frame, i.e. in its diagonal representation

MQDT.kmat — Method

kmat(N::Number, M::fModel, P::Parameters)
kmat(N::Number, M::kModel, P::Parameters)

Function that returns the K matrix in the fragmentation frame.

MQDT.magnetic_matrix_cached — Method

See also magneton

magnetic_matrix_cached(nd, mp, ic, k1::Channels, k2::Channels)

Iterates magneton over Channels, returns a matrix. This function is cached using the LRUCache package.

MQDT.magneton — Method

magneton(nd, mp, ic, q1::fjQuantumNumbers, q2::fjQuantumNumbers)
magneton(nd, mp, ic, q1::jjQuantumNumbers, q2::jjQuantumNumbers)

Returns the reduced magnetic dipole moment (excluding radial contributions from the Rydberg electron)

MQDT.mmat — Method

mmat(N::Number, M::Model, P::Parameters)

Function that returns the M matrix.

MQDT.mroots — Method

mroots(N::Number, M::Model, P::Parameters)
mroots(N1::Number, N2::Number, M::Model, P::Parameters)

Function that returns the roots of the M matrix in the range from N to N+1 or, if provided, from N1 to N2. Root finding algorithm used from the package Roots.

MQDT.multipole_moments — Method

multipole_moment(s1::BasisState, s2::BasisState, P::Paramaters)

Returns the reduced matrix elements for electric dipole, electric quadrupole, magnetic dipole, and electric dipole squared.

MQDT.nu — Method

nu(N::Number, T::Vector, I::Number, R::Number)
nu(N::Number, M::Model, P::Parameters)
nu(N::Vector, M::Model, P::Parameters)

Given a principal quantum number, return the channel-dependent principal quantum numbers, depending on the channel thresholds and the Rydberg constant.

Examples

MQDT.nu(60, [50442.795744, 83967.7, 80835.39, 77504.98, 50443.217463], 50443.217463, 109736.9635066)

# output

5-element Vector{Float64}:
 60.41940064523729
  1.80841488964728
  1.8992315361002043
  2.0125837973909384
 60.0

MQDT.parity — Method

See also unique_parity

parity(T::Channels)

Given a channel representation, this function returns the parity of a each channel based of the orbital angular momenta of the core and the Rydberg electron.

MQDT.radial_matrix — Method

See also radial_moment_cached

function radial_matrix(s1::BasisState, s2::BasisState, order::Int)

Evaluates radial_moment over two BasisStates, returns a matrix, where each entry corresponds to a pair of sqdt states from the two BasisStates.

MQDT.radial_moment_cached — Method

radial_moment_cached(species::Symbol, nu1::Float64, l1::Int, nu2::Float64, l2::Int, order::Int)

Calculate the radial matrix element using the python ryd-numerov package. This function is cached using the LRUCache package.

Examples

radial_moment = MQDT.radial_moment_cached(:Yb174, 70.2667, 1, 71.2638, 2, 1)
isapprox(radial_moment, 1712.16; rtol=1e-5)

# output

true

MQDT.rot — Method

rot(T::Number, C::String, D::Int)
rot(T::Vector, C::Vector{String}, D::Int)
rot(N::Number, M::fModel, P::Parameters)

Function that returns a rotation matrix of dimensions D with rotation angle T. Works for subsequent rotations passed as vectors.

Examples

println(MQDT.rot(0.4, "23", 3))
println(MQDT.rot([0.4, 0.2], ["23", "12"], 3))

# output

[1.0 0.0 0.0; 0.0 0.9210609940028851 -0.3894183423086505; 0.0 0.3894183423086505 0.9210609940028851]
[0.9800665778412416 -0.19866933079506122 0.0; 0.18298657129998708 0.90270109637546 -0.3894183423086505; 0.07736548146578168 0.3816559020950483 0.9210609940028851]

MQDT.square_brakets — Method

square_brakets(a::Vector) = sqrt(prod(2a.+1))

Auxiliary function to calculate angular overlaps (Robicheaux2018 Eq. 11)

Examples

MQDT.square_brakets([1, 2, 3])

# output

10.246950765959598

MQDT.state_data — Method

state_data(T::BasisArray, P::Parameters)
state_data(T::DataBaseArray, P::Parameters)

Store a bound state data as a data frame as used by the PAIRINTERACTION software.

MQDT.theta — Method

See also theta_rr

theta(N::Number, M::Matrix)
theta(N::Vector, M::Matrix)

Given M contains quantum defect parameters, where the k-th column represents the (k-1)st order of energy dependence, this function returns the specific quantum defect at principal quantum number N, where N can also be a list.

Examples

MQDT.theta(60, [0.4 10 100; 0.4 1 0])

# output

2-element Vector{Float64}:
 0.4027854938271605
 0.4002777777777778

MQDT.theta_rr — Method

See also theta

theta_rr(N::Number, M::Matrix, index::Int)

Similar to theta, but here the energy dependence is assumed to be according to the Rydberg-Ritz formula. Given M contains quantum defect parameters, where the k-th column represents the (k-1)st order of energy dependence, this function returns the specific quantum defect at principal quantum number N, where N can also be a list.

Examples

MQDT.theta_rr(60, [0.4 10 100; 0.4 1 0], 1)

# output

0.4028231137913411

MQDT.transform — Method

transform(N::Number, M::fModel, P::Parameters)

Function that returns the frame transformation matrix from close-coupling to fragmentation frame.

MQDT.unique_parity — Method

See also parity

unique_parity(T::Channels)

Given a channel representation, this function checks the parity of each channel and returns it in case each channel has the same parity.

MQDT.Λ — Method

Λ(x) = sqrt((2x+1) * (x+1) * x)

Auxiliary function to calculate reduced matrix elements for the magnetic moment (Robicheaux2018 Eq. 25)

Examples

MQDT.Λ(5)

# output

18.16590212458495